refactor: separate overridden settings from the rest of the logic
(cherry picked from commit 10d8c2b3446d84d3fae7045e7cdd379cecf139c9)
