+++ c8run/internal/overrides/overrides.go
@@ -1,0 +1,42 @@
+/*
+ * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
+ * one or more contributor license agreements. See the NOTICE file distributed
+ * with this work for additional information regarding copyright ownership.
+ * Licensed under the Camunda License 1.0. You may not use this file
+ * except in compliance with the Camunda License 1.0.
+ */
+ package overrides
+
+ import (
+         "fmt"
+         "os"
+ )
+
+func SetEnvVars(javaHome string) error {
+	envVars := map[string]string{
+		"CAMUNDA_OPERATE_CSRFPREVENTIONENABLED":                  "false",
+		"CAMUNDA_OPERATE_IMPORTER_READERBACKOFF":                 "1000",
+		"CAMUNDA_REST_QUERY_ENABLED":                             "true",
+		"CAMUNDA_TASKLIST_CSRFPREVENTIONENABLED":                 "false",
+		"ZEEBE_BROKER_EXPORTERS_ELASTICSEARCH_ARGS_BULK_DELAY":   "1",
+		"ZEEBE_BROKER_EXPORTERS_ELASTICSEARCH_ARGS_BULK_SIZE":    "1",
+		"ZEEBE_BROKER_EXPORTERS_ELASTICSEARCH_ARGS_INDEX_PREFIX": "zeebe-record",
+		"ZEEBE_BROKER_EXPORTERS_ELASTICSEARCH_ARGS_URL":          "http://localhost:9200",
+		"ZEEBE_BROKER_EXPORTERS_ELASTICSEARCH_CLASSNAME":         "io.camunda.zeebe.exporter.ElasticsearchExporter",
+		"ES_JAVA_HOME": javaHome,
+		"ES_JAVA_OPTS": "-Xms1g -Xmx1g",
+	}
+
+	for key, value := range envVars {
+		currentValue := os.Getenv(key)
+		if currentValue != "" {
+			continue
+		}
+		if err := os.Setenv(key, value); err != nil {
+			return fmt.Errorf("failed to set environment variable %s: %w", key, err)
+		}
+	}
+
+	return nil
+}
+
